refactor: relocate dashboard layout components
The components for dashboard layout were moved from standalone
directory into a general layout directory.
