+++ src/components/layouts/dashboard/DashboardSidebar.tsx
@@ -91,6 +91,26 @@ export default function DashboardSidebar({
       ),
     },
     {
+      name: "Messages",
+      icon: (
+        <svg
+          width="18"
+          height="18"
+          viewBox="0 0 18 18"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+          className="stroke-current"
+        >
+          <path
+            d="M9 5.7797V16.613M9 5.7797C9 5.7797 9 4.53572 9 4.11304C9 3.19256 9.74619 2.44637 10.6667 2.44637C11.5871 2.44637 12.3333 3.19256 12.3333 4.11304C12.3333 5.03351 11.5871 5.7797 10.6667 5.7797C10.1689 5.7797 9 5.7797 9 5.7797ZM9 5.7797C9 5.7797 9 4.16546 9 3.69637C9 2.54578 8.06726 1.61304 6.91667 1.61304C5.76607 1.61304 4.83333 2.54578 4.83333 3.69637C4.83333 4.84696 5.76607 5.7797 6.91667 5.7797C7.59817 5.7797 9 5.7797 9 5.7797ZM3.16667 9.11304H14.8333M3.16667 9.11304C2.24619 9.11304 1.5 8.36684 1.5 7.44637C1.5 6.5259 2.24619 5.7797 3.16667 5.7797H14.8333C15.7538 5.7797 16.5 6.5259 16.5 7.44637C16.5 8.36684 15.7538 9.11304 14.8333 9.11304M3.16667 9.11304L3.16667 14.9464C3.16667 15.8668 3.91286 16.613 4.83333 16.613H13.1667C14.0871 16.613 14.8333 15.8668 14.8333 14.9464V9.11304"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+        </svg>
+      ),
+    },
+    {
       name: "Listings",
       icon: (
         <svg
