+++ src/pages/listings/index.tsx
@@ -1,4 +1,4 @@
-import DashboardLayout from "components/dashboard/DashboardLayout";
+import DashboardLayout from "components/layouts/dashboard/DashboardLayout";
 import UserListings from "features/listings/userListings/UserListings";
 import { GetServerSideProps } from "next";
